requirements.txt
aiohttp==3.6.2
asyncpg==0.21.0
discord.py==1.4.1
matplotlib==3.3.1
networkx==2.4
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.4.0.42
pandas==1.1.0
pickleshare==0.7.5
#Pillow==7.2.0
#plotly==4.9.0
requests==2.24.0
requests-oauthlib==1.3.0
scipy==1.5.2
bs4==0.0.1
asyncpgsa==0.26.3
psycopg2-binary==2.8.6
SQLAlchemy==1.3.19
